+++ src/Curiosity.TempFiles/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- Fixed failure when temp directory doesn't exist
+
 ## [1.1.0] - 2021-03-27
 
 ### Added